+++ storybook/stories/_support/StoryDecorator.js
@@ -3,6 +3,7 @@ import React, { Component, Children, Fragment } from 'react';
 import type { Node } from 'react';
 import { observer } from 'mobx-react';
 import { ThemeProvider } from 'react-polymorph/lib/components/ThemeProvider';
+import { SimpleSkins } from 'react-polymorph/lib/skins/simple';
 import { daedalusTheme } from '../../../source/renderer/app/themes/daedalus';
 import { themeOverrides } from '../../../source/renderer/app/themes/overrides';
 
@@ -20,7 +21,11 @@ export default class StoryDecorator extends Component<Props> {
   render() {
     const { children, propsForChildren } = this.props;
     return (
-      <ThemeProvider theme={daedalusTheme} themeOverrides={themeOverrides}>
+      <ThemeProvider
+        theme={daedalusTheme}
+        skins={SimpleSkins}
+        themeOverrides={themeOverrides}
+      >
         <Fragment>
           {Children.map(children, child => {
             const childProps = child.type === 'div' ? {} : { propsForChildren };