+++ source/renderer/app/components/staking/delegation-setup-wizard/DelegationSetupWizardDialog.js
@@ -74,6 +74,7 @@ export default class DelegationSetupWizardDialog extends Component<Props> {
       case 1:
         content = (
           <DelegationStepsChooseWalletDialog
+            numberOfStakePools={stakePoolsList.length}
             stepsList={stepsList}
             wallets={wallets}
             minDelegationFunds={minDelegationFunds}