+++ source/renderer/app/containers/wallet/WalletReceivePage.js
@@ -4,12 +4,16 @@ import { defineMessages, FormattedHTMLMessage } from 'react-intl';
 import { observer, inject } from 'mobx-react';
 import { ellipsis } from '../../utils/strings';
 import WalletReceive from '../../components/wallet/receive/WalletReceive';
+import WalletReceiveDialog from '../../components/wallet/receive/WalletReceiveDialog';
 import VerticalFlexContainer from '../../components/layout/VerticalFlexContainer';
 import NotificationMessage from '../../components/widgets/NotificationMessage';
 import successIcon from '../../assets/images/success-small.inline.svg';
 import type { InjectedProps } from '../../types/injectedPropsType';
+import WalletAddress from '../../domains/WalletAddress';
+import Wallet from '../../domains/Wallet';
 import { ADDRESS_COPY_NOTIFICATION_DURATION } from '../../config/timingConfig';
 import { ADDRESS_COPY_NOTIFICATION_ELLIPSIS } from '../../config/formattingConfig';
+import { generateFileNameWithTimestamp } from '../../../../common/utils/files';
 
 export const messages = defineMessages({
   message: {
@@ -23,6 +27,8 @@ type Props = InjectedProps;
 
 type State = {
   copiedAddress: string,
+  addressToShare?: ?WalletAddress,
+  activeWallet: ?Wallet,
 };
 
 @inject('stores', 'actions')
@@ -32,41 +38,23 @@ export default class WalletReceivePage extends Component<Props, State> {
 
   state = {
     copiedAddress: '',
+    addressToShare: null,
+    activeWallet: this.props.stores.wallets.active,
   };
 
   componentWillUnmount() {
     this.closeNotification();
   }
 
-  closeNotification = () => {
-    const { wallets } = this.props.stores;
-    const wallet = wallets.active;
-    if (wallet) {
-      const notificationId = `${wallet.id}-copyNotification`;
-      this.props.actions.notifications.closeActiveNotification.trigger({
-        id: notificationId,
-      });
-    }
-  };
-
-  render() {
-    const { copiedAddress } = this.state;
-    const { actions } = this.props;
-    const { uiNotifications, wallets, addresses } = this.props.stores;
-    const wallet = wallets.active;
+  get notification() {
+    const { copiedAddress, activeWallet } = this.state;
 
     // Guard against potential null values
-    if (!wallet)
+    if (!activeWallet)
       throw new Error('Active wallet required for WalletReceivePage.');
 
-    const walletAddress = addresses.active ? addresses.active.id : '';
-    const isWalletAddressUsed = addresses.active
-      ? addresses.active.used
-      : false;
-    const walletAddresses = addresses.all.slice().reverse();
-
-    const notification = {
-      id: `${wallet.id}-copyNotification`,
+    return {
+      id: `${activeWallet.id}-copyNotification`,
       duration: ADDRESS_COPY_NOTIFICATION_DURATION,
       message: (
         <FormattedHTMLMessage
@@ -81,37 +69,120 @@ export default class WalletReceivePage extends Component<Props, State> {
         />
       ),
     };
+  }
+
+  closeNotification = () => {
+    const { id } = this.notification;
+    if (id) {
+      this.props.actions.notifications.closeActiveNotification.trigger({ id });
+    }
+  };
+
+  handleIsAddressValid = (index: number) => index < 3 || index > 7;
+
+  handleCopyAddress = (address: string) => {
+    this.setState({ copiedAddress: address });
+    this.props.actions.notifications.open.trigger({
+      id: this.notification.id,
+      duration: this.notification.duration,
+    });
+  };
+
+  handleShareAddress = (addressToShare: WalletAddress) => {
+    this.setState({
+      addressToShare,
+    });
+    const dialog = WalletReceiveDialog;
+    this.props.actions.dialogs.open.trigger({ dialog });
+  };
+
+  handleCloseShareAddress = () => {
+    this.props.actions.dialogs.closeActiveDialog.trigger();
+  };
+
+  handleDownloadPDF = (note: string) => {
+    const { addressToShare } = this.state;
+
+    const name = generateFileNameWithTimestamp({
+      prefix: 'daedalus-cardano-ada-address',
+      extension: '',
+      isUTC: false,
+    });
+
+    // TODO: refactor this direct access to the dialog api
+    const filePath = global.dialog.showSaveDialog({
+      defaultPath: `${name}.pdf`,
+      filters: [
+        {
+          name,
+          extensions: ['pdf'],
+        },
+      ],
+    });
+
+    // if cancel button is clicked or path is empty
+    if (!filePath || !addressToShare) return;
+
+    const { id: address } = addressToShare;
+
+    this.props.actions.wallets.generateAddressPDF.trigger({
+      address,
+      note,
+      filePath,
+    });
+  };
+
+  render() {
+    const {
+      uiNotifications,
+      uiDialogs,
+      addresses,
+      networkStatus,
+      profile,
+      sidebar,
+    } = this.props.stores;
+    const { isIncentivizedTestnet } = networkStatus.environment;
+    const { addressToShare, activeWallet } = this.state;
+    const { currentLocale } = profile;
+    const { isShowingSubMenus } = sidebar;
+
+    // Guard against potential null values
+    if (!activeWallet)
+      throw new Error('Active wallet required for WalletReceivePage.');
+
+    const walletAddresses = addresses.all.slice().reverse();
 
     return (
       <Fragment>
         <VerticalFlexContainer>
           <WalletReceive
-            walletAddress={walletAddress}
-            isWalletAddressUsed={isWalletAddressUsed}
             walletAddresses={walletAddresses}
-            onCopyAddress={address => {
-              this.setState({ copiedAddress: address });
-              actions.notifications.open.trigger({
-                id: notification.id,
-                duration: notification.duration,
-              });
-            }}
+            isAddressValid={this.handleIsAddressValid}
+            onShareAddress={this.handleShareAddress}
+            onCopyAddress={this.handleCopyAddress}
+            isIncentivizedTestnet={isIncentivizedTestnet}
+            isShowingSubMenus={isShowingSubMenus}
+            currentLocale={currentLocale}
           />
         </VerticalFlexContainer>
 
         <NotificationMessage
           icon={successIcon}
-          show={uiNotifications.isOpen(notification.id)}
-          onClose={() => {
-            actions.notifications.closeActiveNotification.trigger({
-              id: notification.id,
-            });
-          }}
+          show={uiNotifications.isOpen(this.notification.id)}
+          onClose={this.closeNotification}
           clickToClose
           hasCloseButton
         >
-          {notification.message}
+          {this.notification.message}
         </NotificationMessage>
+        {uiDialogs.isOpen(WalletReceiveDialog) && addressToShare && (
+          <WalletReceiveDialog
+            address={addressToShare}
+            onCopyAddress={this.handleCopyAddress}
+            onDownloadPDF={this.handleDownloadPDF}
+            onClose={this.handleCloseShareAddress}
+          />
+        )}
       </Fragment>
     );
   }