+++ CHANGELOG.md
@@ -11,10 +11,10 @@ Changelog
 
 ### Chores
 
+- Refactor wallet nav to use new react-polymorph Dropdown ([PR 1593](https://github.com/input-output-hk/daedalus/pull/1593))
 - Fixed build mode of webpack auto dll plugin ([PR 1606](https://github.com/input-output-hk/daedalus/pull/1606))
 - Changed delete wallet button layout for emphasized location/importance and removed export wallet feature ([PR 1612](https://github.com/input-output-hk/daedalus/pull/1612), [PR 1619](https://github.com/input-output-hk/daedalus/pull/1619))
 - Speedup storybook builds in development ([PR 1607](https://github.com/input-output-hk/daedalus/pull/1607))
-- Refactor wallet nav to use new react-polymorph Dropdown ([PR 1593](https://github.com/input-output-hk/daedalus/pull/1593))
 - Added note to UTXO screen showing pending transactions ([PR 1589](https://github.com/input-output-hk/daedalus/pull/1589))
 - Fixed broken source maps ([PR 1594](https://github.com/input-output-hk/daedalus/pull/1594))
 - Reorganized Storybook by domain ([PR 1537](https://github.com/input-output-hk/daedalus/pull/1537))