+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
 
 ### Chores
 
+- Refactor wallet nav to use new react-polymorph Dropdown ([PR 1593](https://github.com/input-output-hk/daedalus/pull/1593))
 - Fixed build mode of webpack auto dll plugin ([PR 1606](https://github.com/input-output-hk/daedalus/pull/1606))
 - Changed delete wallet button layout for emphasized location/importance and removed export wallet feature ([PR 1612](https://github.com/input-output-hk/daedalus/pull/1612), [PR 1619](https://github.com/input-output-hk/daedalus/pull/1619))
 - Speedup storybook builds in development ([PR 1607](https://github.com/input-output-hk/daedalus/pull/1607))